Moved InputChangeController class from SearchController to its separate file.
